// backend/src/types/validation-error.type.ts
export type ValidationErrorType =
    | "required"
    | "invalid_format"
    | "min_length"
    | "enum"
    | "url"
    | "array"
    | "empty"
    | "invalid"
    | "max_length"
    | "min_numeric"
    | "max_numeric"
    | "already_exists"
    | "not_match"
    | "weak_password"
    | "alpha"
    | "numeric"
    | "boolean"
    | "date"
    | "string"
;